CREATES GAME PLAYING FN AND UPDATES GAME FN
This creates the function which will play the game function for 5 times and keep track of the scores of both the computer and the player.

This also updates the playGame function so that the result is just win,loss or tie and the result is generated based on that by including computer choice and player choice. This is done to improve code readability as well.