SETS UP BUTTON LISTENERS
This will set up listener and handler functions in web page so that when user clicks on the button, the game function will be executed, replacing the old way inwhich on loading the player asks for user input. This is done to make the page more user-friendly and simpler to play.